+++ tests/function/test_check_all.py
@@ -43,4 +43,4 @@ def test_missing_space_before_act(function):
 
     assert excinfo.value.line_number == 4
     assert excinfo.value.offset == 4
-    assert excinfo.value.text == 'AAA03 Expected 1 blank line before Act block, found none'
+    assert excinfo.value.text == 'AAA03 expected 1 blank line before Act block, found none'